requirements.txt
pyyaml
osm2geojson